+++ src/parsing.jl
@@ -83,7 +83,7 @@ julia> measurement("-1234e-1")
 """
 measurement(str::AbstractString) = parse(Measurement{Float64}, str)
 
-function Base.tryparse(::Type{Measurement{T}}, str::S) where {T<:AbstractFloat, S<:AbstractString}
+function Base.tryparse(::Type{Measurement{T}}, str::S) where {T<:Real, S<:AbstractString}
     m = match(rxp_error_with_parentheses, str)
     if m !== nothing # "123(45)e6"
         val_str::S, val_dec, err_str::S, err_dec_str, expn = m.captures
@@ -128,7 +128,7 @@ function Base.tryparse(::Type{Measurement{T}}, str::S) where {T<:AbstractFloat, 
     return measurement(val, err)
 end
 
-function Base.parse(::Type{Measurement{T}}, str::S) where {T<:AbstractFloat, S<:AbstractString}
+function Base.parse(::Type{Measurement{T}}, str::S) where {T<:Real, S<:AbstractString}
     out = tryparse(Measurement{T}, str)
     out === nothing && throw(ArgumentError("cannot parse $(repr(str)) as Measurement{$T}"))
     return out